requirements.txt
streamlit
opencv-python-headless
Pillow
numpy
tensorflow
scikit-image
matplotlib
keras
requests
h5py